Added zero padding in frequency space
Address issue #108

The resize_coeff method now takes a scale factor and instead of the size of the
desired array. Currently the scale factors must be integer values. This for of
zero padding call spectral interpolation should persevere the general shape of
the influence coefficients when they are scaled to a larger domain. Currently the influence coefficients can only be scaled by an integer value.