Add Spectral Interpolation notebooks to explore resizing.
Address issue #108

Add Spectral_Interpolation and Spectral_interpolation_influence_coefficinets
notebooks. The results from both of these notebooks are not promising. The
literature mentions that an appropriate placement of zeros in the frequency
space allows for rescaling in the frequency domain, but results from these
notebooks should that the placement difficult.