Update example
address #243

Update example by reducing the size of the images and updating the
overall work flow. More work needs to be done on this example before
it is ready to be added to the development branch. The images in the
example may need to be hand picked for the example.